Applied changes to turret and base class and made tests for these classes

+++ core/src/com/tda367/game/Model/Turret.java
@@ -3,15 +3,15 @@ package com.tda367.game.Model;
 import com.tda367.game.Interfaces.Shoot;
 
 public class Turret implements Shoot {
-    private int damage;
+    private double damage;
     private double shootingSpeed;
 
-    public Turret(int damage, double shootingSpeed){
+    public Turret(double damage, double shootingSpeed){
         this.damage = damage;
         this.shootingSpeed = shootingSpeed;
     }
 
-    public int getDamage() {
+    public double getDamage() {
         return damage;
     }
